rename hook files
